third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-frame-scrolled-expected.txt

Verifies that a tap occurring in a scrolled iframe has the correct co-ordinates

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending GestureTapDown to 30,27
Sending GestureShowPress
Sending GestureTap
Received mousemove in child frame
PASS event.screenX is pointInWindow.x
PASS event.screenY is pointInWindow.y
PASS event.clientX is pointInFrame.x
PASS event.clientY is pointInFrame.y
PASS event.pageX is pointInFrame.x + frameScrollOffset.x
PASS event.pageY is pointInFrame.y + frameScrollOffset.y
Received mousedown in child frame
PASS event.screenX is pointInWindow.x
PASS event.screenY is pointInWindow.y
PASS event.clientX is pointInFrame.x
PASS event.clientY is pointInFrame.y
PASS event.pageX is pointInFrame.x + frameScrollOffset.x
PASS event.pageY is pointInFrame.y + frameScrollOffset.y
Received mouseup in child frame
PASS event.screenX is pointInWindow.x
PASS event.screenY is pointInWindow.y
PASS event.clientX is pointInFrame.x
PASS event.clientY is pointInFrame.y
PASS event.pageX is pointInFrame.x + frameScrollOffset.x
PASS event.pageY is pointInFrame.y + frameScrollOffset.y
Received click in child frame
PASS event.screenX is pointInWindow.x
PASS event.screenY is pointInWindow.y
PASS event.clientX is pointInFrame.x
PASS event.clientY is pointInFrame.y
PASS event.pageX is pointInFrame.x + frameScrollOffset.x
PASS event.pageY is pointInFrame.y + frameScrollOffset.y
PASS eventReceived.mousemove is true
PASS eventReceived.mousedown is true
PASS eventReceived.mouseup is true
PASS eventReceived.click is true
PASS successfullyParsed is true

TEST COMPLETE
